+++ src/main/resources/config.properties
@@ -23,8 +23,8 @@ TIME_TO_START_RECORDING = 4200
 N_SIMS = 1
 # True to write time series for each core indicator (boolean)
 recordCoreIndicators = false
-# True to write micro data for each transaction made (boolean)
-recordMicroData = false
+# True to write data for each transaction (boolean)
+recordTransactions = true
 # True to write agent based data - very storage intensive (boolean)
 recordAgentData = false
 # True to write agents decision data (boolean)
